FAIT-Calculators/OfferShield/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pdf2image==1.17.0
pytesseract==0.3.10
openai==1.14.0
python-dotenv==1.0.1
supabase==2.3.1
stripe==8.0.0
pydantic==2.6.3
